Tweak the code to comply with coding standard
