+++ src/main/java/duke/Duke.java
@@ -3,17 +3,17 @@ package duke;
 import duke.command.Command;
 import duke.exception.DukeException;
 import duke.exception.InvalidCommandException;
-import duke.task.Tasklist;
+import duke.task.TaskList;
 
 public class Duke {
-    private Tasklist tasks;
+    private TaskList tasks;
     private Ui ui;
     private Storage storage;
 
     public Duke() {
         this.ui = new Ui();
         this.storage = new Storage(this.ui);
-        this.tasks = new Tasklist(this.storage.load(), this.ui);
+        this.tasks = new TaskList(this.storage.load(), this.ui);
     }
 
     public void run() {
